+++ src/main/java/co/kr/allpick/domain/member/dto/mypage/PasswordChangeRequestDto.java
@@ -1,4 +1,4 @@
-package co.kr.allpick.domain.member.dto;
+package co.kr.allpick.domain.member.dto.mypage;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
